Fix API signature construction for payload
The API requires that the pre-HMAC digest be of the following format:
(nonce + client_id + api_key), which it was not. This might be an API
change since the last release of this gem.

The HMAC digest also should be using the API secret without an MD5
digest.

With this the client can communicate with private API endpoints again.