Refactor Withdraw and Deposit Address methods to use new Coin class type
To DRY up the way deposit_address and withdraw methods were prefixed
with the coin type, we now pass in the coin type into these methods.

If we want to add more coin types that QuadrigaCX might add, it'll be
easy to just add new coin types.

A new coin type was added to help prevent bad data being passed in,
along with guard statements.

Revamped the tests to check for valid, nil, and invalid coin type usage
for these methods.

Reflected changes in the README as well.